% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{SCE-miscellaneous}
\alias{SCE-miscellaneous}
\alias{show,SpatialCellExperiment-method}
\title{Miscellaneous SpatialCellExperiment methods}
\description{
Miscellaneous methods for the \linkS4class{SpatialCellExperiment} class
that do not fit in any other documentation category.
}
\section{Available methods}{

In the following code snippets, \code{x} and \code{object} are \linkS4class{SpatialCellExperiment} objects.
\describe{
\item{\code{show(object)}:}{Print a message to screen describing the contents of \code{object}.}
}
}

\examples{
example(SpatialCellExperiment, echo=FALSE) # Using the class example

show(sce)

}
\author{
Kevin Rue-Albrecht
}
