% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SpatialCellExperiment.R
\docType{class}
\name{SpatialCellExperiment-class}
\alias{SpatialCellExperiment-class}
\alias{SpatialCellExperiment}
\title{The SingleCellExperiment class}
\usage{
SpatialCellExperiment(..., spatialDim = NULL)
}
\arguments{
\item{...}{Arguments passed to the \link[SingleCellExperiment:SingleCellExperiment-class]{SingleCellExperiment::SingleCellExperiment} constructor to fill the slots of the base class.}

\item{spatialDim}{A matrix of spatial coordinates.}
}
\description{
The SingleCellExperiment class
}
\examples{
ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
v <- log2(u + 1)

coordinates <- matrix(runif(ncells*3), ncells)
colnames(coordinates) <- c("x", "y", "z")

sce <- SpatialCellExperiment(assays=list(counts=u, logcounts=v),
    spatialDim=coordinates)
sce

}
