% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDim.R
\docType{methods}
\name{spatialDim}
\alias{spatialDim}
\alias{spatialDimNames}
\alias{spatialDim,SpatialCellExperiment-method}
\alias{spatialDimNames,SpatialCellExperiment-method}
\alias{spatialDim<-}
\alias{spatialDimNames<-}
\alias{spatialDim<-,SpatialCellExperiment-method}
\alias{spatialDimNames<-,SpatialCellExperiment,character-method}
\title{Reduced dimensions methods}
\description{
Methods to get or set dimensionality reduction results in a \linkS4class{SpatialCellExperiment} object.
These are typically used to store and retrieve spatial representations of single-cell datasets.
Each row of a spatial dimension result is expected to correspond to a column of the SpatialCellExperiment object.
}
\section{Getters}{

In the following examples, \code{x} is a \linkS4class{SpatialCellExperiment} object.
\describe{
\item{\code{spatialDim(x, withDimnames=TRUE)}:}{
Retrieves the matrix (or matrix-like object) containing spatial dimension coordinates for cells (rows) and dimensions (columns).
If \code{withDimnames=TRUE}, row names of the output matrix are replaced with the column names of \code{x}.
}
\item{\code{spatialDimNames(x)}:}{
Returns a character vector containing the colnames the spatial dimensions in \code{x}.
}
}
}

\section{Setter}{

\code{spatialDim(x) <- value} will replace the spatial dimension matrix
in a \linkS4class{SpatialCellExperiment} object \code{x}.

\code{value} is expected to be a matrix or matrix-like object with number of rows equal to \code{ncol(x)}.
}

\section{Other setters}{

In the following examples, \code{x} is a \linkS4class{SpatialCellExperiment} object.
\describe{
\item{\code{spatialDimNames(x) <- value}:}{
Replaces all colnames for the spatial dimension matrix in \code{x} with a character vector \code{value}.
}
}
}

\examples{
example(SpatialCellExperiment, echo=FALSE)

spatialDim(sce)
spatialDimNames(sce)

}
\author{
Kevin Rue-Albrecht
}
